+++ CHANGELOG.md
@@ -7,7 +7,13 @@
 - change API parameter names for POST requests from snake case to lower hyphen ([#398])
 - change indicator and report keys from lower camel case to lower hyphen ([#398])
 - rename API parameter `layerKey` and `layer` to `topic` ([#501])
+- remove GHS POP Comparison indicators: `ghs_pop_comparison_buildings` and `ghs_pop_comparison_roads` ([#515])
 - remove JRC related layers and report `jrc_requirements` ([#503])
+
+### Bug Fixes
+
+- mapping-saturation: add missing edge case detection for too few data points. ([#512])
+- mapping-saturation: allow result value above 100% ([#479])
 
 ### New Features
 
@@ -37,6 +43,7 @@
 
 [#388]: https://github.com/GIScience/ohsome-quality-analyst/pull/388
 [#398]: https://github.com/GIScience/ohsome-quality-analyst/pull/398
+[#479]: https://github.com/GIScience/ohsome-quality-analyst/pull/479
 [#481]: https://github.com/GIScience/ohsome-quality-analyst/pull/481
 [#482]: https://github.com/GIScience/ohsome-quality-analyst/pull/482
 [#483]: https://github.com/GIScience/ohsome-quality-analyst/pull/483
@@ -44,6 +51,8 @@
 [#498]: https://github.com/GIScience/ohsome-quality-analyst/pull/498
 [#501]: https://github.com/GIScience/ohsome-quality-analyst/pull/501
 [#503]: https://github.com/GIScience/ohsome-quality-analyst/pull/503
+[#512]: https://github.com/GIScience/ohsome-quality-analyst/pull/512
+[#515]: https://github.com/GIScience/ohsome-quality-analyst/pull/515
 
 ## 0.14.0